refactor: extract sign_and_publish_group_op to deduplicate governance handlers
Add ContextManager::sign_and_publish_group_op() helper that encapsulates
the governance_preflight → clone datastore/node_client → sign_apply_and_publish
boilerplate pattern shared by 7 handlers.

Simplified handlers:
- set_group_alias: 47 → 21 lines
- set_default_capabilities: 50 → 27 lines
- set_default_visibility: 56 → 32 lines
- update_group_settings: 42 → 22 lines
- set_member_alias: 62 → 35 lines (keeps own-alias check)
- set_member_capabilities: 64 → 37 lines (keeps membership check)
- update_member_role: 83 → 62 lines (keeps last-admin check)

Net: -125 lines, 0 behavioral changes.
