+++ crates/context/src/handlers/set_group_alias.rs
@@ -1,9 +1,8 @@
-use actix::{ActorResponse, Handler, Message, WrapFuture};
+use actix::{ActorResponse, Handler, Message};
 use calimero_context_client::group::SetGroupAliasRequest;
 use calimero_context_client::local_governance::GroupOp;
-use tracing::info;
 
-use crate::{group_store, ContextManager};
+use crate::ContextManager;
 
 impl Handler<SetGroupAliasRequest> for ContextManager {
     type Result = ActorResponse<Self, <SetGroupAliasRequest as Message>::Result>;
@@ -17,34 +16,6 @@ impl Handler<SetGroupAliasRequest> for ContextManager {
         }: SetGroupAliasRequest,
         _ctx: &mut Self::Context,
     ) -> Self::Result {
-        let preflight = match self.governance_preflight(&group_id, requester, true) {
-            Ok(p) => p,
-            Err(err) => return ActorResponse::reply(Err(err)),
-        };
-
-        let datastore = preflight.datastore.clone();
-        let node_client = preflight.node_client.clone();
-        let sk = preflight.signer_sk();
-        let alias_for_log = alias.clone();
-
-        ActorResponse::r#async(
-            async move {
-                group_store::sign_apply_and_publish(
-                    &datastore,
-                    &node_client,
-                    &group_id,
-                    &sk,
-                    GroupOp::GroupAliasSet {
-                        alias: alias.clone(),
-                    },
-                )
-                .await?;
-
-                info!(?group_id, %alias_for_log, "group alias set");
-
-                Ok(())
-            }
-            .into_actor(self),
-        )
+        self.sign_and_publish_group_op(&group_id, requester, true, GroupOp::GroupAliasSet { alias })
     }
 }
