+++ crates/context/src/handlers/set_member_capabilities.rs
@@ -1,10 +1,8 @@
-use actix::{ActorResponse, Handler, Message, WrapFuture};
+use actix::{ActorResponse, Handler, Message};
 use calimero_context_client::group::SetMemberCapabilitiesRequest;
 use calimero_context_client::local_governance::GroupOp;
-use tracing::info;
 
-use crate::group_store;
-use crate::ContextManager;
+use crate::{group_store, ContextManager};
 
 impl Handler<SetMemberCapabilitiesRequest> for ContextManager {
     type Result = ActorResponse<Self, <SetMemberCapabilitiesRequest as Message>::Result>;
@@ -19,11 +17,7 @@ impl Handler<SetMemberCapabilitiesRequest> for ContextManager {
         }: SetMemberCapabilitiesRequest,
         _ctx: &mut Self::Context,
     ) -> Self::Result {
-        let preflight = match self.governance_preflight(&group_id, requester, true) {
-            Ok(p) => p,
-            Err(err) => return ActorResponse::reply(Err(err)),
-        };
-
+        // Verify the target member actually exists in the group
         if group_store::get_group_member_role(&self.datastore, &group_id, &member)
             .ok()
             .flatten()
@@ -34,29 +28,14 @@ impl Handler<SetMemberCapabilitiesRequest> for ContextManager {
             )));
         }
 
-        let datastore = preflight.datastore.clone();
-        let node_client = preflight.node_client.clone();
-        let sk = preflight.signer_sk();
-
-        ActorResponse::r#async(
-            async move {
-                group_store::sign_apply_and_publish(
-                    &datastore,
-                    &node_client,
-                    &group_id,
-                    &sk,
-                    GroupOp::MemberCapabilitySet {
-                        member,
-                        capabilities,
-                    },
-                )
-                .await?;
-
-                info!(?group_id, %member, capabilities, "member capabilities updated");
-
-                Ok(())
-            }
-            .into_actor(self),
+        self.sign_and_publish_group_op(
+            &group_id,
+            requester,
+            true,
+            GroupOp::MemberCapabilitySet {
+                member,
+                capabilities,
+            },
         )
     }
 }
